+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## UNRELEASED
+- Looting an Item Frame is now a crime.
+
 ## 1.2.3 - 2026-05-05
 - [Fabric] Fixed picking the block with CarryOn not being considered as crime.